requirements.txt
wheel
pytorch_lightning
black
ruff
hydra-core>=1.3.2
jaxtyping
beartype
wandb==0.17.6
einops
colorama
scikit-image
colorspacious
matplotlib
moviepy==1.0.3
imageio
dacite
lpips
e3nn
tabulate
svg.py
opencv-python
sk-video
numpy
packaging
ninja
easydict
omegaconf
Pillow
plyfile
PyYAML
tensorboardX
timm
tqdm